This Project is based on replacing non-vegetarian meals with vegetarian meals or green meals. Basically when the user enters in this website then he will know that animals & other living things has also their life, then there is the possibility of they will not see them as a kind of food or meal.
This Project is based on replacing non-vegetarian meals with vegetarian meals or green meals. Basically when the user enters in this website then he will know that animals & other living things has also their life, then there is the possibility of they will not see them as a kind of food or meal.